refactor: move the auth middleware
